# Rust use bump allocator for keys