--- a/rust/optimized/main.rs
+++ b/rust/optimized/main.rs
@@ -6,6 +6,7 @@
 // There's nothing particularly interesting here other than swapping out std's
 // default hashing algorithm for one that isn't cryptographically secure.
 
+use std::cell::Cell;
 use std::error::Error;
 use std::io::{self, BufWriter, Read, Write};
 
@@ -17,7 +18,6 @@ use std::io::{self, BufWriter, Read, Write};
 //
 // N.B. This crate brings in a new hashing function. We still use std's hashmap
 // implementation.
-use bumpalo::Bump;
 use fxhash::FxHashMap as HashMap;
 
 fn main() {
@@ -30,7 +30,7 @@ fn main() {
 fn try_main() -> Result<(), Box<dyn Error>> {
     let stdin = io::stdin();
     let mut stdin = stdin.lock();
-    let keys = Bump::new();
+    let keys = Cell::new(Vec::with_capacity(256 * 1024)); // more than enough
     let mut counts: HashMap<&[u8], u64> = HashMap::default();
     let mut buf = vec![0; 64 * (1 << 10)];
     let mut offset = 0;
@@ -78,7 +78,7 @@ fn try_main() -> Result<(), Box<dyn Error>> {
     Ok(())
 }
 
-fn increment<'bump>(keys: &'bump Bump, counts: &mut HashMap<&'bump [u8], u64>, word: &[u8]) {
+fn increment<'a>(keys_outer: &'a Cell<Vec<u8>>, counts: &mut HashMap<&'a [u8], u64>, word: &[u8]) {
     // using 'counts.entry' would be more idiomatic here, but doing so requires
     // allocating a new Vec<u8> because of its API. Instead, we do two hash
     // lookups, but in the exceptionally common case (we see a word we've
@@ -88,6 +88,11 @@ fn increment<'bump>(keys: &'bump Bump, counts: &mut HashMap<&'bump [u8], u64>, w
         return;
     }
     // store keys in another buffer to avoid unnecessary allocation for each key
-    let word = keys.alloc_slice_copy(word);
+    let mut keys = keys_outer.take();
+    let start = keys.len();
+    keys.extend_from_slice(word);
+    // Safety: extend the lifetime of keys since are reusing the same buffer like an arena
+    let word = unsafe { std::mem::transmute(&keys[start..start + word.len()]) };
     counts.insert(word, 1);
+    keys_outer.set(keys); // store it back
 }