cc @BurntSushi 

Previously #77

Now this is using vec as local keys buffer instead bumpalo.

Before

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      60.7 ms ±  11.8 ms    [User: 53.3 ms, System: 7.5 ms]
  Range (min … max):    43.9 ms …  74.8 ms    44 runs
```

After keys arena

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      47.7 ms ±  11.5 ms    [User: 41.2 ms, System: 6.5 ms]
  Range (min … max):    37.4 ms …  70.0 ms    48 runs
 ```

After bufwriter

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      46.7 ms ±  13.5 ms    [User: 43.4 ms, System: 3.3 ms]
  Range (min … max):    29.9 ms …  63.9 ms    43 runs
```

With only standard library vec (without bumpalo)

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      41.8 ms ±  11.5 ms    [User: 38.6 ms, System: 3.0 ms]
  Range (min … max):    31.0 ms …  64.8 ms    69 runs
```

A single line of unsafe is required to extend the lifetime for keys which we already known. That is what the 3rd party crate does, I don't see any way around that. Even though we can do transmute without unsafe but that is slower (mean ~100ms).