@BurntSushi 2021-03-18
@pickfire Nice! That's a good improvement. Personally, I agree with @benhoyt's "no unsafe" constraint though. Maybe this could be added as an `optimized-unsafe` Rust variant alongside the others instead. Up to @benhoyt though!