@pickfire 2021-03-18
IMHO it isn't optimized without this, how can we keep bunch of allocations while still calling it optimized? :P Arena is one way but that requires third party crate, unsafe is another way, I don't know any other way to avoid the unnecessary allocations without unsafe or arena. I wonder how...