@BurntSushi 2021-03-18
Well, there is also `optimized-customhashmap`. I think reasonable people can disagree. But basically, once you open the `unsafe` flood gates, it's tough to draw a line. On the one hand, people might look at your benchmark and go, "Well, sure, Rust is fast but look, you had to use `unsafe` to do it. Shouldn't Rust be fast without needing `unsafe` all the time?" But put it another way, you might say, "I expect Rust won't be able to be fast in every case without using `unsafe`, so I am interested in seeing how to optimize Rust programs via `unsafe`." Both are valid perspectives IMO. That's why I suggested adding it as a variant.

Anyway, benchmarks are models. All models are wrong. But some are useful. Designing them is an art-form. We could go in circles all day arguing this. At the end of the day, if we have to pick one, I think it's more useful to see how fast Rust can be without `unsafe` in the context of this benchmark.