@benhoyt 2021-03-18
Yeah, I agree with Burntsushi here -- I think this is useful, but I want to let `optimized.rs` stand without "unsafe". Feel free to move to `optimized-unsafe`. I know it is stated a bit loosey goosey in my constraints, "*try* not to use unsafe language features".