@vbarrielle 2021-03-19
Am I right to think this usage of unsafe is only valid if the `Vec`'s capacity is never changed? Maybe there should be such an assertion, otherwise this can be truly unsafe if the input text is in a language with a huge amount of unique words.