@benhoyt 2021-03-19
Per discussion above, I'm going to close this for now. Feel free to reopen or submit a new PR with this as an `optimized-unsafe` version. Thanks anyway!